% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_vr_surv.R
\name{setup_vr_surv}
\alias{setup_vr_surv}
\title{Format raw demographic data for survival modeling}
\usage{
setup_vr_surv(df)
}
\arguments{
\item{df}{A dataframe containing raw demographic data. The dataframe should
contain columns titled "year", "surv_t1" and "area_t0" at minimum.}
}
\value{
A subsetted dataframe of individuals with known survival status.
}
\description{
Format raw demographic data for survival modeling
}
\examples{
x <- data.frame( year = c( 2000, 2000, 2001, 2001, 2002, 2002 ),
                 indiv = c( 1, 2, 1, 2, 1, 2 ),
                 area_t0 = c( 0.002, 0.01, 0.01, 0.05, 0.03, NA ),
                 surv_t1 = c( 1, 1, 1, 1, 0, NA ) )
setup_vr_surv( x )
}
